Исправлены замечания и добавлена проверка текущего фильма в App
