+++ src/components/catalog/catalog-similar.jsx
@@ -3,14 +3,14 @@ import PropTypes from 'prop-types';
 import {filmProps} from "../film-screen/film-screen.prop";
 import Catalog from "./catalog";
 
-const CatalogSmall = (props) => {
+const CatalogSimilar = (props) => {
   return <Catalog className={`catalog--like-this`} header={`More like this`} {...props} />;
 };
 
-Catalog.propTypes = {
+CatalogSimilar.propTypes = {
   films: PropTypes.arrayOf(
       PropTypes.shape(filmProps)
   ),
 };
 
-export default CatalogSmall;
+export default CatalogSimilar;
