+++ src/components/film-screen/film-screen.prop.js
@@ -13,6 +13,7 @@ export const filmProps = {
   description: PropTypes.string.isRequired,
   director: PropTypes.string.isRequired,
   actors: PropTypes.arrayOf(PropTypes.string).isRequired,
+  starring: PropTypes.arrayOf(PropTypes.string).isRequired,
   duration: PropTypes.number.isRequired,
   rating: PropTypes.number.isRequired,
   reviewsCount: PropTypes.number.isRequired,
