+++ src/utils/reviews-panel.js
@@ -1,0 +1,5 @@
+import dayjs from "dayjs";
+
+export const humanizeDate = (date, format) => {
+  return dayjs(date).format(format);
+};
